Seperate out WorkflowList as its own component
